Move synthetic events
